+++ core/dbt/deprecations.py
@@ -57,40 +57,6 @@ class DispatchPackagesDeprecation(DBTDeprecation):
     '''
 
 
-class NotADictionaryDeprecation(DBTDeprecation):
-    _name = 'not-a-dictionary'
-
-    _description = '''\
-    The object ("{obj}") was used as a dictionary. In a future version of dbt
-    this capability will be removed from objects of this type.
-    '''
-
-
-class ModelsKeyNonModelDeprecation(DBTDeprecation):
-    _name = 'models-key-mismatch'
-
-    _description = '''\
-    "{node.name}" is a {node.resource_type} node, but it is specified in
-    the {patch.yaml_key} section of {patch.original_file_path}.
-
-
-
-    To fix this warning, place the `{node.name}` specification under
-    the {expected_key} key instead.
-
-    This warning will become an error in a future release.
-    '''
-
-
-class AdapterMacroDeprecation(DBTDeprecation):
-    _name = 'adapter-macro'
-    _description = '''\
-    The "adapter_macro" macro has been deprecated. Instead, use the
-    `adapter.dispatch` method to find a macro and call the result.
-    adapter_macro was called for: {macro_name}
-    '''
-
-
 class PackageRedirectDeprecation(DBTDeprecation):
     _name = 'package-redirect'
     _description = '''\
@@ -138,9 +104,6 @@ active_deprecations: Set[str] = set()
 
 deprecations_list: List[DBTDeprecation] = [
     DispatchPackagesDeprecation(),
-    NotADictionaryDeprecation(),
-    ModelsKeyNonModelDeprecation(),
-    AdapterMacroDeprecation(),
     PackageRedirectDeprecation()
 ]