+++ test/integration/016_macro_tests/test_macros.py
@@ -95,20 +95,6 @@ class TestAdapterMacroNoDestination(DBTIntegrationTest):
         assert "In dispatch: No macro named 'dispatch_to_nowhere' found" in str(exc.value)
 
 
-class TestDispatchMacroUseParent(DBTIntegrationTest):
-    @property
-    def schema(self):
-        return "test_macros_016"
-
-    @property
-    def models(self):
-        return "dispatch-inheritance-models"
-
-    @use_profile('redshift')
-    def test_redshift_inherited_macro(self):
-        self.run_dbt(['run'])
-
-
 class TestMacroOverrideBuiltin(DBTIntegrationTest):
     @property
     def schema(self):
@@ -124,7 +110,6 @@ class TestMacroOverrideBuiltin(DBTIntegrationTest):
             'config-version': 2,
             'macro-paths': ['override-get-columns-macros'],
         }
-
 
     @use_profile('postgres')
     def test_postgres_overrides(self):
@@ -164,3 +149,28 @@ class TestDispatchMacroOverrideBuiltin(TestMacroOverrideBuiltin):
     def test_postgres_overrides(self):
         self.run_dbt(["deps"])
         super().test_postgres_overrides()
+
+
+class TestAdapterMacroDeprecated(DBTIntegrationTest):
+
+    @property
+    def schema(self):
+        return "test_macros_016"
+
+    @property
+    def models(self):
+        return "deprecated-adapter-macro-models"
+
+    @property
+    def project_config(self):
+        return {
+            'config-version': 2,
+            "macro-paths": ["deprecated-adapter-macro"]
+        }
+
+    @use_profile('postgres')
+    def test_postgres_invalid_macro(self):
+        with pytest.raises(dbt.exceptions.CompilationException) as exc:
+            self.run_dbt(['run'])
+
+        assert 'The "adapter_macro" macro has been deprecated' in str(exc.value)